requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
greenlet==2.0.2
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wordcloud==1.9.2
